2024 day 13 part 2 wip

+++ src/main/java/fr/ninauve/renaud/adventofcode/year2024/day13/Location.java
@@ -1,6 +1,6 @@
 package fr.ninauve.renaud.adventofcode.year2024.day13;
 
-public record Location(int x, int y) {
+public record Location(long x, long y) {
 
     public Location add(Location delta) {
         return new Location(x+delta.x, y+delta.y);
